deleted old files (collaborator_parallel  and old versions of manager and collaborator). Renamed manager and collaborator
